refactor(windows): 🔨 separate client group policy rule classes
